requirements.txt
argon2-cffi==21.3.0
pytest==7.1.1
